refactor app file and routes file
